Add files via upload
This is a House Prediction Project based on using Eval ML Library. Eval ML created the best pipeline. We used XGBoost Regressor with the best hyperparameters to find out the output of House Price with an accuracy of 81.6 %. We used R2 Score to measure the performance of the model. Below are the results of the model we chose on test data. Output displays the House Price.